+++ sentry_sdk/integrations/pydantic_ai/spans/utils.py
@@ -3,48 +3,16 @@
 from typing import TYPE_CHECKING
 
 import sentry_sdk
-from sentry_sdk._types import BLOB_DATA_SUBSTITUTE
-from sentry_sdk.ai.consts import DATA_URL_BASE64_REGEX
-from sentry_sdk.ai.utils import _set_span_data_attribute, get_modality_from_mime_type
-from sentry_sdk.consts import SPANDATA
+from sentry_sdk.ai.monitoring import record_token_usage
+
+from .._extract import extract_usage
 
 if TYPE_CHECKING:
-    from typing import Any, Dict, Union
+    from typing import Union
 
     from pydantic_ai.usage import RequestUsage, RunUsage
 
     from sentry_sdk.traces import StreamedSpan
-
-
-def _serialize_image_url_item(item: "Any") -> "Dict[str, Any]":
-    """Serialize an ImageUrl content item for span data.
-
-    For data URLs containing base64-encoded images, the content is redacted.
-    For regular HTTP URLs, the URL string is preserved.
-    """
-    url = str(item.url)
-    data_url_match = DATA_URL_BASE64_REGEX.match(url)
-
-    if data_url_match:
-        return {
-            "type": "image",
-            "content": BLOB_DATA_SUBSTITUTE,
-        }
-
-    return {
-        "type": "image",
-        "content": url,
-    }
-
-
-def _serialize_binary_content_item(item: "Any") -> "Dict[str, Any]":
-    """Serialize a BinaryContent item for span data, redacting the blob data."""
-    return {
-        "type": "blob",
-        "modality": get_modality_from_mime_type(item.media_type),
-        "mime_type": item.media_type,
-        "content": BLOB_DATA_SUBSTITUTE,
-    }
 
 
 def _set_usage_data(
@@ -60,34 +28,15 @@ def _set_usage_data(
         span: The Sentry span to set data on.
         usage: RequestUsage or RunUsage object containing token usage information.
     """
-    if usage is None:
+    usage_info = extract_usage(usage)
+    if usage_info is None:
         return
 
-    if hasattr(usage, "input_tokens") and usage.input_tokens is not None:
-        _set_span_data_attribute(
-            span, SPANDATA.GEN_AI_USAGE_INPUT_TOKENS, usage.input_tokens
-        )
-
-    # Pydantic AI uses cache_read_tokens (not input_tokens_cached)
-    if hasattr(usage, "cache_read_tokens") and usage.cache_read_tokens is not None:
-        _set_span_data_attribute(
-            span, SPANDATA.GEN_AI_USAGE_INPUT_TOKENS_CACHED, usage.cache_read_tokens
-        )
-
-    # Pydantic AI uses cache_write_tokens (not input_tokens_cache_write)
-    if hasattr(usage, "cache_write_tokens") and usage.cache_write_tokens is not None:
-        _set_span_data_attribute(
-            span,
-            SPANDATA.GEN_AI_USAGE_INPUT_TOKENS_CACHE_WRITE,
-            usage.cache_write_tokens,
-        )
-
-    if hasattr(usage, "output_tokens") and usage.output_tokens is not None:
-        _set_span_data_attribute(
-            span, SPANDATA.GEN_AI_USAGE_OUTPUT_TOKENS, usage.output_tokens
-        )
-
-    if hasattr(usage, "total_tokens") and usage.total_tokens is not None:
-        _set_span_data_attribute(
-            span, SPANDATA.GEN_AI_USAGE_TOTAL_TOKENS, usage.total_tokens
-        )
+    record_token_usage(
+        span,
+        input_tokens=usage_info.input_tokens,
+        input_tokens_cached=usage_info.cache_read_tokens,
+        input_tokens_cache_write=usage_info.cache_write_tokens,
+        output_tokens=usage_info.output_tokens,
+        total_tokens=usage_info.total_tokens,
+    )
